Tooling for calculating 3x3x3 cube orders
Tooling for calculating 3x3x3 cube orders in a local and distributed fashion.
The simple-client and simple-server implementations are proof of concept. They
have have zero security, and are not scalable beyond a few tens or hundreds of
clients. The C++ implementations of Client and Server are meant to be a more
secure distributed solver implementation, suitable for use by the general
public and scalable to thousands or even hundreds of thousands of clients.

The simple implementation is still beta, but has been tested and appears to be
working. The C++ Client and Server are still very much in work.

Note: This branch is being squashed constantly. Any pulls should take that into
account. You have been warned.